refactor(push-notifications): including push notifications as hook
